Move fragments into view
