+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:background="@color/white"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".frontend.fragments.HomeFragment">
+    tools:context=".frontend.view.HomeFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
